QMath

QBase.QMath — Module

A QBase.jl submodule providing general purpose mathematics useful to quantum mechanics.

Exports:

Matrices:

partial_trace - The partial trace matrix operation.

computational_basis_vectors - A general orthonormal set of vectors.

Matrices

QBase.QMath.computational_basis_vectors — Function

computational_basis_vectors( dim::Int64 ) ::Vector{ Vector{Int64} }

Returns an orthonormal set of column vectors spanning a vector space of dimension dim.

QBase.QMath.partial_trace — Function

partial_trace(
    ρ::AbstractMatrix, subsystem_dims::Vector{Int64}, subsystem_id::Int64
) :: AbstractMatrix

Performs the partial trace on matrix ρ with respect to the subsystem at the specified subsystem_id. The partial trace is a quantum operation $\mathcal{E}$ mapping the input Hilbert space to the output Hilbert space, $\mathcal{E} \; : \; \mathcal{H}_X \rightarrow \mathcal{H}_Y$. A quantum operation admits the operator sum representation, $\mathcal{E}(\rho) = \sum_i E_i \rho E_i^{\dagger}$. For example, given a 3-qubit system with density matrix $\rho_{ABC}$, the partial trace with respect to system $B$, is explicitly,

\[\rho_{AC} = \text{Tr}_B[\rho_{ABC}] = \mathcal{E}_B(\rho_{ABC}) = \sum_i E_i \rho_{ABC} E_i^{\dagger}\]

where $E_i = \mathbb{I}_A \otimes \langle i |_B \otimes\mathbb{I}_C$, represents the Kraus operators for the quantum operation and $\rho_{AC}$ is the reduced density operator remaining after system $B$ is traced out.

Inputs:

ρ : The matrix on which the partial trace is performed. This matrix should be square and have dimension equal to the product of the subsystem_dims.

subsystem_dims : A vector containing positive integer elements which specify the dimension of each subsystem. E.g. A 3-qubit system has subsystem_dims = [2,2,2], a system containing a qubit and a qutrit has subsystem_dims = [2,3].

subsystem_id : A positive integer indexing the subsystem_dims vector to signify the subsytem which the partial trace will trace out.

Output:

The reduced matrix which results after the partial trace is performed on ρ. This matrix is square and has dimension of product of subsystem_dims with the element corresponding to subsystem_id removed.

A DomainError is thrown if ρ is not square or of proper dimension.

Probability