Evolution

QBase.evolve — Function

Apply a unitary evolution U to density matrix ρ: $\rho' = U\rho U^{\dagger}$.

evolve(
    U :: Unitary,
    ρ :: State
) :: State

Apply a unitary evolution U to a state ket ψ: $|\psi'\rangle = U|\psi\rangle$.

evolve(
    U ::  Unitary,
    ψ :: Ket
) :: Ket

Base.:* — Method

*( U :: Unitary, ket :: Ket ) :: Ket
*( bra :: Bra, U :: Unitary ) :: Bra
*(U1 :: Unitary, ρ :: State, U2 ::  Unitary) :: Matrix

Unitaries

QBase.Unitary — Type

Unitary( U :: AbstractMatrix ) <: Operator{T}

Unitary matrices represent the physical evolution of quantum states. The Constructor, Unitary(U), throws a DomainError if the provided matrix, U is not unitary.

QBase.is_unitary — Function

is_unitary( U :: Matrix; atol=ATOL :: Float64 ) :: Bool

Returns true if matrix U is unitary. The hermitian adjoint of a unitary matrix is its inverse:

U' * U == I where I is the identity matrix.

Unitary Operations

Base.:* — Method

*(unitaries :: Vararg{Unitary}; atol=ATOL :: Float64) :: Unitary

Base.kron — Method

kron(unitaries :: Vararg{Unitary}; atol=ATOL :: Float64) :: Unitary

Base.adjoint — Method

adjoint(U :: Unitary) :: Unitary

Unitary Constructors

QBase.σI — Constant

σI :: Unitary{Int64}

Identity unitary:

julia> σI
2×2 Unitary{Int64}:
 1  0
 0  1

QBase.σx — Constant

σx :: Unitary{Int64}

Pauli-X unitary:

julia> σx
2×2 Unitary{Int64}:
 0  1
 1  0

QBase.σy — Constant

σy :: Unitary{Complex{Int64}}

Pauli-Y unitary:

julia> σy
2×2 Unitary{Complex{Int64}}:
 0+0im  0-1im
 0+1im  0+0im

QBase.σz — Constant

σz :: Unitary{Int64}

Pauli-Z unitary:

julia> σz
2×2 Unitary:
 1   0
 0  -1

QBase.qubit_rotation — Function

qubit_rotation( θ :: Real; axis="x" :: String ) :: Unitary{Complex{Float64}}

Returns a unitary which performs a qubit rotation along bloch sphere. θ designates the angle of rotation and axis ("x", "y", "z") designates the cartesian axis about which the qubit is rotated.

QBase.random_unitary — Function

random_unitary( d :: Int64 ) :: Unitary{Complex{Float64}}

Constructs a d x d random unitary matrix according to the Haar measure.

Channels

QBase.replacer_channel — Function

replacer_channel(
    ρ :: State,
    σ :: State,
    μ :: Real
) :: State

The replacer channel replaces the quantum state ρ with quantum state σ with probability μ. The replacer channel is expressed:

\[ \mathcal{R}_{\mu}(\rho) = \mu \rho + (1-\mu)\sigma\]

A DomainError is thrown if

μ does not satisfy 1 ≥ μ ≥ 0

ρ and σ are not the same size

QBase.depolarizing_channel — Function

depolarizing_channel( ρ :: State, μ :: Real ) :: State

The depolarizing channel mixes uniform classical noise into a quantum state ρ. The argument μ describes the amount of noise mixed into the quantum states. For a quantum state $\rho$, the depolarizing channel is expressed:

\[ \mathcal{D}_{\mu}(\rho) = \mu \rho + \frac{(1 - \mu)}{d}\mathbb{I}_d\]

A DomainError is thrown if μ does not satisfy 1 ≥ μ ≥ 0.

QBase.erasure_channel — Function

erasure_channel( ρ :: State, μ :: Real ) :: State

The erasure channel mixes a quantum state ρ with an error flag $|F\rangle$ orthogonal to the Hilbert space of ρ. The argument μ describes the probability that ρ is replaced with the error flag. For a quantum state $\rho$, the erasure channel is expressed:

\[ \mathcal{E}_{\mu}(\rho) = \mu \rho + (1 - \mu) |F \rangle \langle F|\]

Note that the erasure channel increases the dimension of the Hilbert space by 1.

A DomainError is thrown if μ does not satisfy 1 ≥ μ ≥ 0.