QBase.jl - Exports

QBase — Module

A collection of methods and submodules useful for computation of quantum sytems.

Exports

Methods:

evolve - Performs unitary evolution of quantum states.

measurement_probs - Outcome probabilities of quantum measurement.

Modules:

States - Types and constructors for representing quantum states.

Observables - Types and constructors for representing measureable quantities.

Unitaries - Types and constructors for representing unitary operators.

Information - Functions for computing information-theoretic quantities.

QMath - Mathematics useful for modeling quantum operations.

Methods

QBase.evolve — Function

Apply a unitary evolution U to density matrix ρ.

evolve(
    U::Unitaries.AbstractUnitary,
    ρ::States.AbstractDensityMatrix
) :: DensityMatrix

Apply a unitary evolution U to a state ket ψ.

evolve(
    U::Unitaries.AbstractUnitary,
    ψ::States.AbstractKet
) :: Ket

QBase.measurement_probs — Function

Computes the outcome probabilities for a quantum measurement. The conditional probabilities are determined by the Born rule, $P(i|j) = \text{Tr}[\Pi_i \rho_j]$, where $\Pi_j$ is a POVM element and $\rho_j$ is a density matrix.

Measurement of a single Ket or DensityMatrix:

measurement_probs(
    Π :: Observables.AbstractPOVM,
    ρ :: States.AbstractDensityMatrix
) :: QMath.Conditionals

measurement_probs(
    Π :: Observables.AbstractPOVM,
    ψ :: States.AbstractKet
) :: QMath.Conditionals

Measurement of an ensemble of Ket or DensityMatrix types:

measurement_probs(
    Π :: Observables.AbstractPOVM,
    ρ_states :: Vector{<:States.AbstractDensityMatrix}
) :: QMath.Conditionals

measurement_probs(
    Π :: Observables.AbstractPOVM,
    ψ_kets :: Vector{<:States.ABstractKet}
) :: QMath.Conditionals