Operators

QBase.Operator — Type

abstract type Operator{T<:Number} <: AbstractMatrix{Number} end

A matrix representing a linear operator that acts upon a complex-valued Hilbert space. The Operator is an abstract type that parents all linear operator in quantum mechanics. These matrix types can represent quantum states, evolution, and measurements, each with their individual constraint. These constraints are place upon the children of th Operator abstract type.

Operator Algebra

Base.:* — Method

Operator Multiplication:

*(operators :: Vararg{Operator}) :: Matrix

Base.:* — Method

Operator types can multiply Bra and Ket types.

$O|\psi\rangle = | \psi' \rangle$:

*(O :: Operator, ket :: Ket) :: Vector

$\langle \psi |O = \langle \psi'|$:

*(bra :: Bra, O :: Operator) :: Adjoint{T, Vector{T}} where T <: Number

Inner product, $\langle \psi |O|\sigma\rangle = \langle \psi|\sigma'\rangle$:

*(bra :: Bra, O :: Operator, ket :: Ket) :: Number

Base.kron — Method

Kronecker Product:

kron(operators :: Vararg{Operator}) :: Matrix

LinearAlgebra.rank — Method

Matrix Rank:

rank(O :: Operator) :: Int64

Base.sqrt — Method

Matrix Square Root:

sqrt(O :: Operator) :: Matrix