Information

QBase.Information — Module

Methods for quantifying information and randomness.

Entropy quantifiers are taken with respect to base-2 logarithms. The entropy is understood as the number of bits {0,1} required to communicate a random result with certainty.

Entropy

QBase.Information.shannon_entropy — Function

shannon_entropy( probabilities :: QMath.Marginals ) :: Float64

shannon_entropy( probabilities :: Vector{<:Real}) :: Float64

The classical entropy of a probability distribution.

QBase.Information.von_neumann_entropy — Function

von_neumann_entropy( ρ :: States.AbstractDensityMatrix ) :: Float64

von_neumann_entropy( ρ :: Matrix{<:Number} ) :: Float64

The von neumann entropy of a density matrix.

QBase.Information.joint_entropy — Function

joint_entropy(priors :: QMath.Marginals, conditionals :: QMath.Conditionals) :: Float64

Returns the entropy for the union of pdf $P(x,y)$.

QBase.Information.conditional_entropy — Function

conditional_entropy(priors::QMath.Marginals, conditionals::QMath.Conditionals) :: Float64

Returns the conditional entropy for the system with specified priors and conditionals.

Information

QBase.Information.holevo_bound — Function

holevo_bound(
    priors :: QMath.Marginals,
    ρ_states :: Vector{<:AbstractDensityMatrix}
) :: Float64

holevo_bound(
    priors :: Vector{<:Real},
    ρ_states :: Vector{Matrix{<:Number}}
) :: Float64

Computes the upper bound of a quantum channel's information capacity. The information shared through a quantum channel cannot exceed a classical channel of the same dimension.

QBase.Information.holevo_information — Function

holevo_information(
    priors :: QMath.Marginals,
    ρ_states :: Vector{<:AbstractDensityMatrix},
    Π :: Observables.AbstractPOVM
) :: Float64

Computes the holevo (mutual) information shared through a quantum channel.

QBase.Information.mutual_information — Function

mutual_information(
    priors :: QMath.Marginals,
    conditionals :: QMath.Conditionals
) :: Float64

The entropy of the overlap between p(x) and p(y). The information shared from y to x.

State Discrimination

QBase.Information.success_probability — Function

success_probability(
    priors::QMath.Marginals,
    ρ_states::Vector{<:States.AbstractDensityMatrix},
    Π::Observables.AbstractPOVM
) :: Float64

The probability of correctly distinguishing quantum states with the specifed POVM.

QBase.Information.error_probability — Function

error_probability(
    priors::QMath.Marginals,
    ρ_states::Vector{<:States.AbstractDensityMatrix},
    Π::Observables.AbstractPOVM
) :: Float64

The probability of incorrectly distinguishing quantum states with the specifed POVM.